[VSPC-171] pr comments
