+++ vspace/src/main/java/edu/asu/diging/vspace/web/staff/ListSpacesController.java
@@ -43,24 +43,4 @@ public class ListSpacesController {
 
         return "staff/spaces/spacelist";
     }
-    
-    @RequestMapping("/staff/spaces/search")
-    public ResponseEntity<String> listSpacesForSlideContentBlock(@RequestParam(value = "term", required = false) String searchTerm){
-        List<ISpace> spaces = null;
-        if (searchTerm != null && !searchTerm.trim().isEmpty()) {
-            spaces = spaceManager.findByName(searchTerm);
-        } else {
-            spaces = spaceManager.getSpaces(1);
-        }
-        
-        ObjectMapper mapper = new ObjectMapper();
-        ArrayNode idArray = mapper.createArrayNode();
-        for (ISpace space : spaces) {
-            ObjectNode imageNode = mapper.createObjectNode();
-            imageNode.put("id", space.getId());
-            imageNode.put("name", space.getName());
-            idArray.add(imageNode);
-        }
-        return new ResponseEntity<String>(idArray.toString(), HttpStatus.OK);
-    }
 }
