+++ vspace/src/main/webapp/WEB-INF/views/staff/modules/slides/slide.html
@@ -74,11 +74,12 @@
             
             function getSelectedSpace(blockId){
             	$.ajax({
-                	url: "[(@{'/staff/module/'+${module.id}+'/slide/'+${slide.id}+'/spaceblock/'})]"+ blockId + "?" +[[${_csrf.parameterName}]]+'='+[[${_csrf.token}]],
+                	url: "[(@{'/staff/module/'+${module.id}+'/slide/'+${slide.id}+'/space/'})]"+ blockId + "?" +[[${_csrf.parameterName}]]+'='+[[${_csrf.token}]],
     		type: 'GET',
     		success: function(result) {
-    				let selectedSpace = document.getElementById("selectedSpace");
-    				selectedSpace.innerHTML = result;
+    			console.log(result +'hh');
+    			$("#selectedSpace").text('');
+    			$("#selectedSpace").text(result);
     		    	},
     		error: function(data) {
     			var alert = $('<div class="alert alert-danger alert-dismissible fade show" role="alert"><p>We are sorry but something went wrong. Please try to delete again later.</p><button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
@@ -886,7 +887,7 @@
                     formData.append('spaceId', spaceId);
                     markDown.value('');
                     $.ajax({
-                        url: "[(@{'/staff/module/'+${module.id}+'/slide/'+${slide.id}+'/SpaceBlockContent'})]",
+                        url: "[(@{'/staff/module/'+${module.id}+'/slide/'+${slide.id}+'/space'})]",
                         type: 'POST',
                         cache       : false,
                         contentType : false,
